move markers to public folder
