+++ __tests__/components/Markers/TimelineMarkers.test.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { render } from 'react-testing-library'
 import 'jest-dom/extend-expect'
-import TimelineMarkers from 'lib/markers/TimelineMarkers'
-import TodayMarker from 'lib/markers/TodayMarker'
+import TimelineMarkers from 'lib/markers/public/TimelineMarkers'
+import TodayMarker from 'lib/markers/public/TodayMarker'
 import { RenderWrapper } from 'test-utility/marker-renderer'
 
 describe('TimelineMarkers', () => {
