+++ src/lib/markers/public/CursorMarker.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { TimelineMarkersConsumer } from './TimelineMarkersContext'
-import { TimelineMarkerType } from './markerType'
+import { TimelineMarkersConsumer } from '../TimelineMarkersContext'
+import { TimelineMarkerType } from '../markerType'
 
 class CursorMarker extends React.Component {
   static propTypes = {
